requirements.txt
absl-py==1.4.0
accelerate==0.26.1
aiodns==3.1.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
APScheduler==3.10.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
boto3==1.26.151
botocore==1.29.152
brotlipy==0.7.0
bypy==1.8.4
cachetools==4.2.4
cchardet==2.1.7
certifi==2021.10.8
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cityhash==0.2.4.post11
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.4
comm==0.1.3
conda==23.1.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
databricks-cli==0.17.7
datasets==2.12.0
dbpool==1.2.1
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.9.4
defusedxml==0.7.1
-e git+https://github.com/huggingface/diffusers.git@de9c72d58c3eb4bcd50901d0c23c7c46f0cec29b#egg=diffusers
dill==0.3.6
docker==6.1.3
dotmap==1.3.30
dsc-auth==0.1.16
editdistance==0.8.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
ffmpeg-python==0.2.0
filelock==3.12.1
fire==0.5.0
Flask==2.3.2
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.6.0
func-timeout==4.3.5
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==1.34.1
google-api-python-client==2.118.0
google-auth==2.20.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==1.5.0
google-cloud-texttospeech==2.16.1
google-cloud-translate==2.0.1
googleapis-common-protos==1.62.0
greenlet==2.0.2
grpcio==1.55.0
grpcio-reflection==1.48.2
grpcio-status==1.48.2
grpcio-tools==1.48.2
gunicorn==20.1.0
hiredis==2.3.2
hjson==3.1.0
httplib2==0.22.0
huggingface-hub==0.20.3
hydra-core==1.3.2
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.6.0
importlib-resources==5.12.0
infra-component==1.4.4
infra-framework==1.17.5
infra-kconf==1.1.1
infra-kess==1.1.1
infra-keycenter==1.1.1
infra-storage==1.3.1
iniconfig==2.0.0
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kafka-python==2.0.2
kazoo==2.10.0
kiwisolver==1.4.4
ks-accelerate==0.19.0
ks-kafka-python==2.0.3
ks-mlflow==1.0.14
-e git+ssh://git@git.corp.kuaishou.com/mmu-aicreation/common.git@dc230b5bda26366ee9d11ccc2621c8d17f210a91#egg=ksac
less==0.0.1
lit==16.0.5.post0
lmdb==1.4.1
lxml==4.9.4
lz4==3.1.10
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
mysql-connector-python==8.0.31
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
peft==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
pluggy==1.4.0
prettytable==2.5.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.4.0
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
pysmhasher==0.2.5
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pytest==8.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-snappy==0.6.1
pytorch-fid==0.3.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
querystring-parser==1.2.4
redis==4.6.0
regex==2023.6.3
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
sentencepiece==0.1.99
setuptools-scm==7.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminado==0.17.1
threadpool==1.3.2
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets==5.9.0
transformers==4.30.1
triad==0.9.5
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
unpaddedbase64==2.1.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1673575502006/work
-e git+ssh://git@git.corp.kuaishou.com/mmu-vcg/vcg-train/vcg_diffusers_train.git@059a23246b438e053b7b6c5e74947e3100f714ca#egg=vcg_diffusers_train
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
Werkzeug==2.3.6
wget==3.2
widgetsnbextension==4.0.7
xformers==0.0.20
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
